requirements.txt
Flask==1.0.2
gunicorn==19.9.0
google-cloud-vision==0.35.2
google-cloud-storage==1.13.2
google-cloud-datastore==1.7.3
google-cloud-error-reporting==0.30.1
google-cloud-logging==1.10.0
flask-cors